Add TacMap Chat, sync v3 hardening, site rebuild, and release tooling
- TacMap Chat v1 on iOS and Android (crypto, store, wire format, UI)
- Unit sync: presence quality/retention, relay endpoint policy, websocket
  transport, reconnect backoff, inbound frame policy
- Export/import: pipeline, identity journal and commit coordinator
- Security ADRs 002-004, iOS export compliance, threat model updates
- Site: source rebuild with security policy and tests
- Store assets and release verification scripts

+++ android/app/src/main/java/com/tacmap/calibration/WebMercatorTiles.kt
@@ -75,6 +75,8 @@ object WebMercatorTiles {
     }
 
     data class TileRange(val z: Int, val minX: Int, val maxX: Int, val minY: Int, val maxY: Int) {
-        val count: Int get() = (maxX - minX + 1).coerceAtLeast(0) * (maxY - minY + 1).coerceAtLeast(0)
+        val count: Long get() =
+            (maxX - minX + 1).coerceAtLeast(0).toLong() *
+                (maxY - minY + 1).coerceAtLeast(0).toLong()
     }
 }
